Add VSOP87 planetary positions, clean up barrel file and publishing setup
- Implement full planetposition module with VSOP87B data for all 8 planets
- Add VSOP87B coefficient data files (auto-generated from astronomia.js)
- Add LICENSE (MIT), proper README, CHANGELOG
- Simplify barrel file to non-conflicting foundational exports only
- Fix all lint warnings (unused imports, variables, fields)
- Fix math.pow return type in semidiameter and stellar
- Add jdeToJulianYear to julian module
- Update pubspec with topics, homepage; update repo URL
- 182 tests passing

+++ lib/src/semidiameter/semidiameter.dart
@@ -28,7 +28,7 @@ double semidiameter(double s0, double delta) {
 
 /// Asteroid diameter in km from absolute magnitude [h] and albedo [a].
 double asteroidDiameter(double h, double a) {
-  return math.pow(10, 3.12 - 0.2 * h - 0.5 * math.log(a) / math.ln10);
+  return math.pow(10, 3.12 - 0.2 * h - 0.5 * math.log(a) / math.ln10).toDouble();
 }
 
 /// Asteroid apparent semidiameter in radians from diameter [d] (km)
